+++ src/components/ProjectDetails/ProjectDetails.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../../redux/mapStoreToProps';
-import TaskList from '../TaskList';
+import mapStoreToProps from '../../redux/mapStoreToProps';
+import TaskList from '../TaskList/TaskList';
 import './ProjectDetails.css'
