+++ src/components/TaskList/TaskList.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import mapStoreToProps from '../../redux/mapStoreToProps';
 import AddNewTask from '../AddNewTask/AddNewTask';
-import AdminTaskList from './AdminTaskList/AdminTaskList'
-import ContractorTaskList from './ContractorTaskList'
-import ClientTaskList from './ClientTaskList'
+import AdminTaskList from '../AdminTaskList/AdminTaskList'
+import ContractorTaskList from '../Project/ContractorTaskList'
+import ClientTaskList from '../ClientTaskList/ClientTaskList'
 
 const getCookie = (cookieName) => {
   // Get name followed by anything except a semicolon
